Improve the readability of dynamic labels which scroll.
If the text of a dynamic label to be displayed is wider than the given
length, it will be scrolled. Currently such a label starts scrolling
immediately after it is placed and - even more unpleasant - the start of the
text is randomly somewhere within the specified space of the label. Both
makes it hard to track and to read.

Now such a dynamic label starts left-aligned and begins scrolling through the
specified space only after a short delay (2.5 seconds). Every time the start
of the text nears the left margin again during the scrolling process it will
stop and everything starts all over again, i.e. scrolling after a short
delay.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@32848 b3059339-0415-0410-9bf9-f77b7e298cf2