screenshot: make screenshot filenames configurable
This adds the --screenshot-template option, which specifies a template
for the filename used for a screenshot. The '%' character is parsed as
format specifier. These format specifiers insert metadata into the
filename. For example, '%f' is replaced with the filename of the
currently played file.

This uses localtime instead of localtime_r. Unfortunately, mingw doesn't
support localtime_r, and using localtime is better than adding #ifdefs.