screenshot: make png writer use compression setting
Use the compression setting that was originally introduced for the jpg
writer. The input value will be rescaled to the value range the
libavcodec png codec accepts.

It should be taken care that the default compression setting does not
map to the maximum png compression, as that will be very slow even on
fast computers. On the other hand, the lowest compression setting
produces files of several MB size, which should be avoided as well.