Workaround cygwin issue on gethostbyname2 check
Apparently cygwin1.dll does export gethostbyname2 (I don't know if
it's the same "gethostbyname2" mplayer tries to use) , but it isn't
declared in any header in the system. For some reason the configure
test passes despite that but then stream/tcp.c fails with undeclared
function when it is attempted to be compiled. Do some unusual
assignment in the configure test to get the compiler to notice that
dereferencing an implicit "int" from an undeclared function is not a
good idea.