ass: set PlayResY=720 for text subs
Why? When specifying a font size with *Fontsize in ass-force-style,
this font size should be always the same, regardless whether SRT or
ASS subs (with override) are used. Since the override code in libass
assumes a vertical resolution of 720, do the same on the mplayer side
when rendering SRT with libass. 720 is an arbitrary value, but both
sides have to use the same for consistency.