gl_common, vo_gl3: add proper OpenGL 3.x context creation
Create a proper OpenGL 3.x context. It uses the core profile (compatibility
features removed), and is forward-compatible (no deprecated functionality
included).

The function to create the 3.x context is named create_window_gl3, and
replaces the create_window and setGlWindow functions. The separation
between the two old functions doesn't seem to make much sense, and the new
function creates both a window and the GL context.

create_window_gl3 is implemented for X11 only, and vo_gl3 will only run on
X11.

Some uses of legacy functionality in vo_gl3 had to be fixed.

Do not set the GL_TEXTURE_PRIORITY texture parameter for textures in
general. It was removed in modern OpenGL, and likely had no effect anyway.

Remove some unused OpenGL functions from the loader.

Use the proper name for glGetProgramivARB. glGetProgramiv is a different
function and is used for GLSL.