screenshot: sanitize screenshot filenames
Now the filename is cleaned up from characters that are not allowed in
filenames. Any illegal characters are simply replaced with '_'. This is
useful, because strings produced by the screenshot filename template are
not always valid filenames. Consider date & time formats, or the case
when %f is used and the filename is actually a URL with special characters
in it.

Normal strings in the actual screenshot template are intentionally not
sanitized, because it's supposed to be possible to set a screenshot
directory by passing a path, and the path separators must be kept.