vo_gl: add highly experimental support for OpenGL ES
Use glLoadMatrixf. This makes it easier to support OpenGL ES.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34425 b3059339-0415-0410-9bf9-f77b7e298cf2

Add highly experimental support for OpenGL ES.

It only supports EGL/X11, uses/supports only ES v1,
will crash if certain features are used, compiling
without desktop GL installed is not tested and
possibly more caveats.
However it is close enough to be able to display
a video on a BeagleBoard via OpenGL.
Performance could not be tested properly since I do
not have a display that is compatible with the
BeagleBoard output...