vo_corevideo: restructure this video output
Restructure this video output to be similar to vo_gl, even if simpler and less
feature complete (for example it's still missing EOSD support). Ideally, it
should act as a decent fallback in the case where something breaks in the
OSX support of vo_gl.

Here's a summary of what changed:

 * Clean up the code to conform a bit more to the mplayer2 conventions. Enforce
   80 column wrapping, use a private struct for file variables, use the new
   libvo api.

 * Remove the shared buffer code since it wasn't using any function from the
   CoreVideo API. Moreover, it's presence in vo_corevideo was forcing the
   non-GUI related code to perform more image copies than necessary.

 * Add OSD rendering using OpenGL instead of writing directly on the video
   image data.

 * Simplify the logic for the rendering function when dealing with panscan.

 * Add redraw frame support.

 * Add colormatrix support by using the built-in API provided by CoreVideo.