vo_gl3: remove doublebuffer stuff and simplify (E)OSD code
The previous code took different code paths for rendering (especially OSD)
when the -nodouble option is used. The man page says -nodouble is for
debugging purposes, and the different codepath probably wasn't actually
useful for anything. I see no reason why the code should be more
complicated for a useless anti-feature, so remove it.