x11: set mplayer.png as window icon
This will set etc/mplayer.png as _NET_WM_ICON. The icon is embedded in
the mplayer binary. The script bin_to_header.py is added, which is used
to convert the PNG file to a header file containing its contents as byte
array. Add a rule to the Makefile to do this conversion automatically as
part of the build process.

This adds Python (2.x or 3.x) as build dependency.

etc/mplayer.png is copied from svn commit 33572, which says the png was
derived from http://en.wikipedia.org/wiki/File:MPlayer.svg