cocoa_common: accept window drag event from the whole window surface
Change the window to accept mouse drag events not only on the title
bar, but also on the rest of the window surface; this includes the video
area.

It looks like the changing of the window mask resets the behaviour
specified in the delegate method, probably due to some strange interaction
with NSBorderlessWindow. For this reason call -setPresentationOptions
in the -fullscreen method to remind cocoa the behaviour we want.