screenshot: add mode to restart counter for %n filename template
Add a '#' modifer to '%n', which makes the counter restart every time a
screenshot is taken. This can be useful if there is other information in
the screenshot filename (such as the playback time) and %n is only wanted
to make the filename unique.

Valid uses of %n now include for example '%#n' and '%#05n'.