input, x11: always pass through multimedia keys
There are countless of special key symbols defined for special keys,
which are found on so called "multimedia" or "internet" keyboards. These
extra keys should be usable from mplayer, but defining separate "portable"
key codes and adding all the key names seems unfeasible.

As a simpler solution, specify a range of key codes as reserved for these
multimedia keys, and map it directly to the XFree86 specific extension key
code range. Using the XFree86 mappings is justified by the fact that
mplayer is native on X, and that XFree86/Xorg are the most popular X
implementations. On other operating systems, lookup tables will be needed
to do the mapping.

This means that pressing an unbound multimedia key will make mplayer print
the numeric mplayer key code, which users can use to bind the key in their
input.conf.