af: fix crash when trying to use softvol with AC3 pass-through
Changing the volume when softvol is enabled causes insertion of the
"volume" filter, which failed, because it doesn't support AC3. Since the
filter wasn't removed after that, and the filter chain was in a bogus
state, random crashes occured past this point.

Fix it by reinitializing the filter chain completely on failure.

This also affects balance controls.