libmenu: use OSD rendering, and remove old menu video filter code
The menu code used a rather fragile method of drawing the menus: it
rendered directly into the video being displayed, by inserting a menu
specific video filter (vf_menu). There are various disadvantages and
bugs related to this.

Switch rendering to use the OSD code. We render to a screen-sized OSD
object, so not many changes are required. We have to do alpha blending
in C code. Ideally, rendering would be done via libass/EOSD, but for
now, this is good enough.