libmenu: fix crashes
Random parts of mplayer2 have been changed to use talloc instead of
standard malloc/free. This hasn't been done very carefully: the code
in libmenu still assumes it can free a string returned by
mp_property_print with free(), when it should be using talloc_free().
Likewise, get_path() returns a malloc'ed string, but we assign it to
a variable that's managed as m_option_t, which uses talloc.

libmenu was still part of the code when these changes were made.

Also note that while mp_property_print uses talloc, this is not the case
with property_expand_string(). This requires us to add ugly special
casing in menu_param.c.