tmp: cocoa: core: keep only ccooa gui on the main thread
The Cocoa Framework is designed to have the main thread of the
application dedicated to the Cocoa's Runloop. Change mplayer's main() to
spawn a new thread just after it's start and keep the main thread for
the runloop.

Any method that calls to AppKit modifying windows and views is scheduled
on the main thread Runloop since the AppKit classes are not thread safe
and any concurrent access to them causes hard crashes.