Update Libav API uses
Change various code to use the latest Libav API. Because Libav dropped
APIs that were necessary with all Libav versions until quite recently
(like setting avctx->age), this commit drops compatibility with all
but the latest Libav versions. The new APIs used in this commit
existed in the recent Libav release 0.8; however, because of bugs in
the APIs in that release it's not properly compatible. Hopefully the
relevant fixes will be added in a Libav point release.

The libavcodec error_recognition setting has been removed and replaced
with different semantics. I removed the "--lavdopts=er=<value>" option
accordingly, as I don't think it's widely enough used to be worth
attempting to emulate the old option semantics using the new API.
A new option with the new semantics can be added later if needed.