Basic multiuser support
MPlayer now tests to see if a "%appdata%\mplayer" folder exists. If
it exists, then mplayer uses it instead of the "(.exe dir)\mplayer"
for reading/writing config files.

This new behavior is of course only triggered if the user makes a
mplayer" folder inside "%appdata%"; not making it maintains current
behavior.

This isn't the ideal support, but doing a better method would involve
too much coding so I settled for this ;)