windows support: unicode filenames
Windows uses a legacy codepage for char* / runtime functions accepting
char*. Using UTF-8 as codepage with setlocale() is explicitly
forbidden.

Work this around by overriding the MSVCRT functions with custom
wrappers, that assume UTF-8 and use "proper" API calls like _wopen etc.
to deal with unicode filenames.

This is not perfectly clean, but still somewhat sane, and much better
than littering the rest of the mplayer code with MinGW specific hacks.
It's also a bit fragile, but that's actually little different from the
previous situation. Also, MinGW is unlikely to ever include a nice way
of dealing with this.

Replace uses of stat(). We can't "redirect" stat() due to complications
with MinGW.