vo_gl3: add option for gamma correct scaling
Normally, vo_gl3 (and vo_gl) scale the image in the YUV colorspace, and
the conversion to RGB is done last. This is blatantly incorrect:
scaling should happen only in linear RGB.

Add a "srgb" suboption, which is supposed to do gamma-correct scaling.
The video is first converted to a video-sized RGB texture using a FBO.
The resulting texture is then rendered on the screen using scaling.
The RGB texture is in linear RGB, and we use OpenGL's sRGB framebuffer
feature to convert the result back to sRGB.

Note that we intentionally don't make use of sRGB textures to convert
the R'G'B' to linear RGB. In theory, the BT.709 colorspace uses a gamma
of 2.35, which is quite different from sRGB. So we have to do that
manually.

Since 2.35 looks too different from what users would expect, 2.2 is
currently used instead. (See mp_csp_gamma().)

If the input colorspace is RGB, the input is assumed to be in sRGB.

Note that the luma scaler suboption selects the scaler for the converted
RGB texture, i.e. basically it will actually affect both luma
and chroma. The chroma scaler suboption selects the scaler that is used
to interpolate the chroma, since the chroma plane must be scaled to
luma size in order to do the RGB conversion.

The intermediate RGB texture is always requested in 16 bit precision in
order not to lose precision.

The "indirect" suboption can be used instead of "srgb" to enable
conversion to RGB before scaling without using sRGB and gamma
correction.