vo_gl3: change gamma handling with 'srgb'
As suggested by nand`: do scaling in 2.2 gamma, and correct the gamma
to video gamma in the final stage (but before conversion to sRGB). The
video gamma is 2.2, or 2.35 for BT.709 (changed again compared to last
commit).

The video gamma is the same gamma value as it can be controlled with
the 'gamma' slave property, so the user can adjust this according to
the lighting conditions when viewing.

Not sure if it's correct/correctly implemented.