screenshot: allow using properties in screenshot filenames
The screenshot format '%{prop}' can now be used to insert the value of
the property 'prop' in the screenshot filename.

A fallback can be optionally specified. '%{prop:foo}' will literally
insert 'foo' in the filename, if the property 'prop' doesn't exist or is
not available. (This is similar, but not the same as property expansion
in the slave command osd_show_property_text or the --playing-msg
option.)

It is assumed that property names never contain the characters '}' or
':', and that the fallback never contains '}'.

Note that some filename template format strings are redundant to using
properties, such as '%f' and '%{filename}'.