vo_gl3: 2-pass filtering for separable scalers
This has to be enabled with the "scale-sep" suboption. It only works
for the convolution filters.

2-pass filtering is only done when the luma scaler (lscale) is a
convolution filter. Chroma scaling is still done with a single pass
filter. Since the "scale-sep" option implies the "indirect" option,
scaling the video to screen size is done in RGB. The chroma scaler is
only used to undo the chroma subsampling, and thus less important.

Note that the GLSL shader relies even more in the shader compiler
heavily optimizing away redundant operations. This is hopefully no
problem with sane OpenGL implementations.