vo_gl3: add fbo-format suboption, which selects FBO texture format
This allows selecting the internal texture format of the FBOs used with
the "indirect" and "scale-sep" options. It can have influence on image
quality, performance, and memory use.

Note that OpenGL implementations are free to use an actually different
internal format from what the user requested.

If this option turns out to be pointless, I'll remove it again.