cocoa: fix pixel format creation on systems older than 10.7
Change the code to require the OpenGL profile version explicitely only
if we are running on a operating system that is new enough (OpenGL
version is changed with the OS minor version on OSX).

OpenGL implementations from older systems refuse to work with the
undefined constants.

Leave the ifdef'd defines of the constants, so it's possible to cross
compile with 10.6 and 10.7 support using the older SDK and still
use OpenGL3 when the code is executed on 10.7 and better.