vo_gl3: fix downscaling with convolution filters
Nice downscaling basically requires averaging an area of pixels. With
convolution based filters (like lscale=lanczos2 etc.) this can be
achieved by extending the size of the filter by the scale factor. This
requires changing the filter size on window size changes, which makes
this change quite messy.

This has the strange consequence that downscaling might be slower than
upscaling, because downscaling uses larger filter matrices.

I'm not sure whether this is correct, but at least it looks better, and
it shouldn't affect the upscaling code path much.

This behavior can be disabled with the "no-fancy-downscaling"
suboption.

Note that downscaling of video that doesn't use square pixels will be
slightly incorrect. In theory we would have to use different lookup
tables for horizontal and vertical scaling, but the current code was
written with the assumption that these are always the same (this is
still true for upscaling). If horizontal and vertical scaling are
different, we use the larger scale factor, which will result in a
blurry (and incorrect) image in the other direction. This special case
could be easily fixed in exchange for more (and possibly slower) code,
but since I'm lazy I'll just hope it doesn't matter and nobody will try
to watch anamorphic video downscaled.

Some of the initialization code is reorganized to make reinitialization
of the scalers easier. Using the "scale-sep" suboption now only enables
rendering over a FBO indirection if "lscale" uses a convolution filter.