vo_gl3: enable scale-sep by default
This is usually faster than using the simple version of the convolution
filters, even though it goes through 2 FBO indirections. Sometimes it's
slower, but not by much.

Note that this does not enable the other indirection steps if no
convolution filter is used with "lscale".

The "no-scale-sep" suboption can be used to disable this behavior.