vo_gl3: change how scaler suboptions are handled
This will make it error in the suboption parser if a scaler is not
supported, pointing out where the bogus argument is. Also makes it
easier to change scalers during runtime (although that's unused yet).

Unfortunately -vo gl3:lscale=help won't work anymore. I blame the
suboption parser, which will hopefully soon be improved to support
choices (including help output) and so on.