windows support: change how unicode filename support is done
Use a less dirty trick to make unicode filenames work on Windows.

Instead of overriding the functions in question on the linker level,
do it with macros. All code that uses standard functions that take
or return filenames must include osdep/io.h.

Note that stat() can't be overridden, because this causes issues with
the type "struct stat" and the fact that MinGW-w64 defines "stat" as
macro. mplayer code should use mp_stat().