Fix bug with playlist playback.
When <stop> is pressed and you shift to another track and press <play>,
the track is played twice (before advancing to the next element in the
list).

Reported by drake ch, drake.ch hotmail com.

When shifting to another track, set information that at the end of
playback of the current track the next track to be played shall not be
determined (!uiGotoTheNext) only if currently playing. That is because
the track shifting function itself (which already has determined the
next track to be played) will end the playback of the current track (and
start the next one) only if currently playing.

If not playing, the next track to be played after the end of playback of
the current track (determined by the track shifting function) must be
determined again (uiGotoTheNext).


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@34778 b3059339-0415-0410-9bf9-f77b7e298cf2