vo_gl3: add simple ordered dithering
If the video input bit depth is larger than the output bit depth, do
ordered dithering.

The output bit depth is detected automatically (see previous commit),
and if it's unknown, 8 bit are assumed.

If color management is used, the input bit depth is assumed to be
16 bits.

Dithering can be disabled with "-vo gl3:dither-depth=-1".