stream_cdda: print CDTEXT if available
Requires compilation with libcdio. The per-CD info will be printed on
playback start, per-track info when a track is played. (This is not a
technical restriction, and just goes along with the code that was there
before.)

The following fields are not included in output, because these are
apparently binary: CDTEXT_DISCID, CDTEXT_GENRE, CDTEXT_SIZE_INFO,
CDTEXT_TOC_INFO, CDTEXT_TOC_INFO2.