configure: enable x11 opengl backend only if cocoa is disabled
The recommended way to get function pointers to the functions in the
OpenGL library is through dlopen/dlsym/dlclose. This causes problems in
the Cocoa OpenGL backend when -lGL (x11's OpenGL headers) is linked
to the binary together with -framework OpenGL.
The linked OpenGL symbols are always from -lGL, causing all the function
pointers to point to null when getFunctions is called against a Cocoa
OpenGL context.

For this reason change the configure autodetection code to disable x11's
gl backend when cocoa is active.